moved app directory to src folder
